Criado popup de confirmacao para exclusao de imagem

+++ src/app/components/galeria/galeria.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { MatGridListModule } from '@angular/material/grid-list';
 import { CommonModule } from '@angular/common';
 import { FotosService } from '../../services/fotos.service';
-import { ConfigService } from '../../services/config.service'; // Importando o ConfigService
+import { ConfigService } from '../../services/config.service';
 import { DomSanitizer, SafeUrl } from '@angular/platform-browser';
 import { Imagem } from '../../models/imagem';
 import { forkJoin, map } from 'rxjs';
@@ -17,18 +17,19 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 })
 export class GaleriaComponent implements OnInit {
   isPopupOpen = false;
+  isDeleteConfirmationOpen = false;
   selectedImageSrc: SafeUrl | null = null;
   selectedQrcodeSrc: SafeUrl | null = null;
-  companyQrcodeSrc: SafeUrl | null = null; // QR code da empresa
-  showCompanyQrcode = false; // Controla exibição do QR code da empresa
+  companyQrcodeSrc: SafeUrl | null = null;
+  showCompanyQrcode = false;
   imagens: SafeUrl[] = [];
   imagensIds: number[] = [];
   fotoAtualId: number | null = null;
-  private qrcodeTimeout: any; // Para gerenciar o timeout
+  private qrcodeTimeout: any;
 
   constructor(
     private fotosService: FotosService,
-    private configService: ConfigService, // Injetando ConfigService
+    private configService: ConfigService,
     private sanitizer: DomSanitizer,
     private snackBar: MatSnackBar
   ) {}
@@ -62,20 +63,17 @@ export class GaleriaComponent implements OnInit {
     this.isPopupOpen = true;
     console.log('ID da Foto:', fotoId);
 
-    // Carregar QR code da foto
     this.fotosService.getQrcodeById(fotoId).subscribe((qrcodeBlob: Blob) => {
       const objectURL = URL.createObjectURL(qrcodeBlob);
       this.selectedQrcodeSrc = this.sanitizer.bypassSecurityTrustUrl(objectURL);
       console.log('ID do QR Code:', fotoId);
     });
 
-    // Carregar QR code da empresa
     this.configService.getQrCode().subscribe((companyQrcodeBlob: Blob) => {
       const objectURL = URL.createObjectURL(companyQrcodeBlob);
       this.companyQrcodeSrc = this.sanitizer.bypassSecurityTrustUrl(objectURL);
-      this.showCompanyQrcode = true; // Mostrar QR code da empresa
+      this.showCompanyQrcode = true;
 
-      // Após 5 segundos, ocultar QR code da empresa
       this.qrcodeTimeout = setTimeout(() => {
         this.showCompanyQrcode = false;
       }, 5000);
@@ -86,25 +84,37 @@ export class GaleriaComponent implements OnInit {
     this.isPopupOpen = false;
     this.selectedImageSrc = null;
     this.selectedQrcodeSrc = null;
-    this.companyQrcodeSrc = null; // Limpar QR code da empresa
+    this.companyQrcodeSrc = null;
     this.showCompanyQrcode = false;
     this.fotoAtualId = null;
     if (this.qrcodeTimeout) {
-      clearTimeout(this.qrcodeTimeout); // Cancelar timeout se fechar o popup
+      clearTimeout(this.qrcodeTimeout);
     }
   }
 
-  deleteImagePopup(): void {
-    this.fotosService.delete(this.fotoAtualId).subscribe(() => {
-      this.snackBar.open('Imagem excluída com sucesso!', 'Fechar', {
-        duration: 5000,
+  openDeleteConfirmationPopup(): void {
+    this.isDeleteConfirmationOpen = true;
+  }
+
+  closeDeleteConfirmationPopup(): void {
+    this.isDeleteConfirmationOpen = false;
+  }
+
+  confirmDelete(): void {
+    if (this.fotoAtualId !== null) {
+      this.fotosService.delete(this.fotoAtualId).subscribe(() => {
+        this.snackBar.open('Imagem excluída com sucesso!', 'Fechar', {
+          duration: 5000,
+        });
+        this.closeDeleteConfirmationPopup();
+        this.closeImagePopup();
+        this.carregarImagens();
+      }, ex => {
+        this.snackBar.open('Erro ao excluir a imagem!', 'Fechar', {
+          duration: 5000,
+        });
+        this.closeDeleteConfirmationPopup();
       });
-      this.closeImagePopup();
-      this.carregarImagens();
-    }, ex => {
-      this.snackBar.open('Erro ao excluir a imagem!', 'Fechar', {
-        duration: 5000,
-      });
-    });
+    }
   }
 }
